+++ src/lib/components/Form/GenericForm.svelte
@@ -56,7 +56,6 @@
 		},
 		...options
 	});
-
 	const { form, enhance } = superform;
 
 	setContext('superform', superform);