+++ CHANGELOG.md
@@ -60,6 +60,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   [#76](https://github.com/thejoltjoker/kreacon/pull/76)
 - Add i18n report script to check for missing translations
   [#78](https://github.com/thejoltjoker/kreacon/pull/78)
+- Allow user to update username on profile page
+  [#82](https://github.com/thejoltjoker/kreacon/pull/82)
 
 ### Changed